Stop using optimize=True when saving Deep Zoom tiles
optimize mode is slower for PNG and reportedly significantly slower for
JPEG, so the dynamic tiler shouldn't be using it.  optimize mode also
disables streaming encoding in libjpeg, so with sufficiently large tiles
we start getting buffer overrun errors.  We could expand the encode buffer
(based on tile size) or enable optimization in the static tiler only for
PNG, but without a clear upside we'll just disable the option.

Closes #1.