tests: call os.add_dll_directory() on Windows
On Python >= 3.8 we need to use os.add_dll_directory() to add OpenSlide
to the DLL path.  There doesn't seem to be a way to arrange that from the
pytest command line, so add some glue code to the tests package that
reads the path from the OPENSLIDE_PATH env var.