Configure pytest to use importlib import mode
By default, running `pytest` prepends the source directory to the Python
path, since that's where the `tests` package is.  That causes Python to
import `openslide` from the source directory instead of using the
installed copy, and this fails because the source directory doesn't have
the compiled extension module.  Configure pytest not to change the Python
path, as recommended by pytest docs:

https://docs.pytest.org/en/7.1.x/explanation/goodpractices.html#choosing-an-import-mode

Then, manually set the Python path to allow tests to load common code.

Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>