Add AbstractSlide.color_profile property
For any given slide, every image returned by read_region() will have the
same color profile.  It's important that we attach it to every region to
document the correct color space, but applications that want to do color
conversion may find it more efficient to build a transform once, up front.
Add a property to help them do this.

Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>