Add ICC profile in _AssociatedImageMap
If an associated image has a profile and it's equal to the one from the
main image, reuse the latter to avoid keeping multiple copies in RAM.

Add a synthetic DICOM fixture so we can use OpenSlide's DICOM driver to
test this.  The fixture was created by hand-hacking an ICC profile into
small.svs and then converting it to DICOM with
com.pixelmed.convert.TIFFToDicom.

For https://github.com/openslide/openslide/pull/477.

Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>