convert: use Python Limited API on Python 3.11 and above
This allows us to produce forward-compatible wheels and stop building
new wheels for every Python minor release.  The buffer protocol API was
added to the Limited API in 3.11, so we still need version-specific wheels
for releases older than that.

Replace a couple functions that aren't in the Stable API with equivalents
that are.

Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>