Support openslide-bin
Prefer loading OpenSlide from the openslide-bin package if available.

For ease of use, document openslide-bin as the preferred installation
option.  Don't document manual installation of openslide-bin binaries on
Linux and macOS, since there isn't generally an advantage to doing that,
but leave the existing Windows Zip instructions for compatibility and to
explain os.add_dll_directory().

Add an explanatory ModuleNotFoundError on Linux if OpenSlide is not found.

Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>